+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- If a single page is being used for multiple products the generated feed now makes sure a unique `identifier` (URL) and `productId` is being used.
+
 ## [1.3.0] - 2025-04-22
 ### Added
 - `get_sc_post_types` filter to manipulate the post types where the meta box will be shown.